requirements.txt
imageio==2.23.0
ipython==8.7.0
matplotlib==3.5.2
numpy>=1.17
Pillow==9.3.0
tensorflow==2.6.0
